requirements.txt
pillow
pillow-heif
piexif
